requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
ruff==0.12.7
sqlparse==0.5.3
typing_extensions==4.14.1
